Class divided into packages
