+++ src/main/java/Deadline.java
@@ -1,3 +1,5 @@
+package duke.task;
+
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
@@ -5,12 +7,12 @@ import java.time.format.DateTimeFormatter;
 public class Deadline extends Task {
     LocalDate day;
 
-    Deadline(String type, String detail, boolean marked, LocalDate day){
+    public Deadline(String type, String detail, boolean marked, LocalDate day){
         super(type, detail, marked);
         this.day = day;
     }
 
-    Deadline(String type, String detail, String day){
+    public Deadline(String type, String detail, String day){
         super(type, detail);
         this.day = LocalDate.parse(day);
     }
