+++ src/main/java/Event.java
@@ -1,3 +1,5 @@
+package duke.task;
+
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
@@ -6,13 +8,13 @@ public class Event extends Task {
     LocalDate start;
     LocalDate end;
 
-    Event(String type, String detail, boolean marked, String start, String end){
+    public Event(String type, String detail, boolean marked, String start, String end){
         super(type, detail, marked);
         this.start = LocalDate.parse(start);
         this.end = LocalDate.parse(end);
     }
 
-    Event(String type, String detail, String start, String end){
+    public Event(String type, String detail, String start, String end){
         super(type, detail);
         this.start = LocalDate.parse(start);
         this.end = LocalDate.parse(end);
